requirements.txt
pydantic[dotenv]~=1.8.2
loguru~=0.5.3
aiorun~=2021.8.1
python-dotenv~=0.19.0
aiosocks~=0.2.6
dpkt~=1.9.7.2